src/org/totalbeginner/tutorial/Book.java
package org.totalbeginner.tutorial;

public class Book {

	String title;
	String author;
	Person person;

	public Book(String string) {
		this.title = string;
		this.author = "unknown author";
		//Person randomPerson = new Person("unknown person");
		//this.setPerson(randomPerson);
	}
	
	/**
	 * @return the author
	 */
	public String getAuthor() {
		return author;
	}

	/**
	 * @param author the author to set
	 */
	public void setAuthor(String author) {
		this.author = author;
	}

	/**
	 * @return the title
	 */
	public String getTitle() {
		return title;
	}

	public void setPerson(Person p2) {
		this.person = p2;
	}

	public Person getPerson() {
		return this.person;
	}
	
	public String toString() {
		return this.getTitle();
	}

}
